feat: migrated appView.js to ts
